refactor: facade에서 event가 처리되지 않도록 로직 분리
- event의 경우 *EventHandler를 통해 처리하도록 수정
