+++ apps/commerce-api/src/main/java/com/loopers/application/payment/PaymentFacade.java
@@ -22,11 +22,8 @@ import com.loopers.domain.payment.PaymentService;
 import com.loopers.domain.point.Point;
 import com.loopers.domain.point.PointAccountEntity;
 import com.loopers.domain.point.PointService;
-import com.loopers.domain.product.ProductService;
-import com.loopers.domain.push.PushService;
 import com.loopers.domain.shared.DomainEvent;
 import com.loopers.domain.shared.DomainEventPublisher;
-import com.loopers.domain.user.UserService;
 import com.loopers.domain.vo.Price;
 import com.loopers.support.error.CoreException;
 import com.loopers.support.error.ErrorType;
@@ -49,9 +46,6 @@ public class PaymentFacade {
 	private final CouponService couponService;
 	private final CouponDiscountCalculator couponDiscountCalculator;
 	private final DomainEventPublisher eventPublisher;
-	private final ProductService productService;
-	private final PushService pushService;
-	private final UserService userService;
 
 	@Transactional
 	public PaymentInfo paymentByPoint(PointPaymentCommand command) {
@@ -141,49 +135,5 @@ public class PaymentFacade {
 		if (status == TransactionStatus.FAIL) {
 			eventPublisher.publish(DomainEvent.of(PaymentEvent.PaymentFail.of(payment.getOrderId(), payment.getId())));
 		}
-	}
-
-	@Transactional
-	public void paymentSuccess(PaymentEvent.PaymentSuccess event) {
-		OrderEntity order = orderService.getOrder(event.orderId());
-		order.paymentConfirm();
-
-		PaymentEntity payment = paymentService.getById(event.paymentId());
-		payment.success();
-
-		var user = userService.getUser(order.getUserId())
-				.orElseThrow(() -> new CoreException(ErrorType.NOT_FOUND, "사용자를 찾을 수 없습니다."));
-
-		pushService.sendOrderSuccessPush(
-				new PushService.UserInfo(
-						user.getLoginId(),
-						user.getName()
-				),
-				new PushService.OrderInfo(
-						order.getOrderStatus().name(),
-						order.getOrderItems()
-								.stream()
-								.map(item -> new PushService.OrderItemInfo(
-										productService.getProduct(item.getProductId()).getName(),
-										item.getQuantity()
-								))
-								.toList()
-				),
-				new PushService.PaymentInfo(
-						payment.getPaymentMethod().name(),
-						payment.getAmount()
-				)
-		);
-	}
-
-	@Transactional
-	public void paymentFail(PaymentEvent.PaymentFail event) {
-		OrderEntity order = orderService.getOrder(event.orderId());
-		order.paymentFailed();
-
-		PaymentEntity payment = paymentService.getById(event.paymentId());
-		payment.fail();
-
-		order.getOrderItems().forEach(item -> productService.increaseStock(item.getProductId(), item.getQuantity()));
 	}
 }
