+++ apps/commerce-api/src/main/java/com/loopers/interfaces/listener/coupon/CouponListener.java
@@ -1,6 +1,6 @@
 package com.loopers.interfaces.listener.coupon;
 
-import com.loopers.application.coupon.CouponFacade;
+import com.loopers.application.coupon.CouponEventHandler;
 import com.loopers.domain.payment.PaymentEvent;
 import lombok.RequiredArgsConstructor;
 import org.springframework.scheduling.annotation.Async;
@@ -12,11 +12,11 @@ import org.springframework.transaction.event.TransactionalEventListener;
 @Component
 public class CouponListener {
 
-	private final CouponFacade couponFacade;
+	private final CouponEventHandler couponEventHandler;
 
 	@Async
 	@TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
 	public void handlePaymentSuccessEvent(PaymentEvent.PaymentSuccess event) {
-		couponFacade.useCoupon(event);
+		couponEventHandler.useCoupon(event);
 	}
 }
