+++ apps/commerce-api/src/main/java/com/loopers/interfaces/listener/payment/PaymentListener.java
@@ -1,6 +1,6 @@
 package com.loopers.interfaces.listener.payment;
 
-import com.loopers.application.payment.PaymentFacade;
+import com.loopers.application.payment.PaymentEventHandler;
 import com.loopers.domain.payment.PaymentEvent;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.event.EventListener;
@@ -9,15 +9,15 @@ import org.springframework.stereotype.Component;
 @RequiredArgsConstructor
 @Component
 public class PaymentListener {
-	private final PaymentFacade paymentFacade;
+	private final PaymentEventHandler paymentEventHandler;
 
 	@EventListener
 	public void handlePaymentSuccessEvent(PaymentEvent.PaymentSuccess event) {
-		paymentFacade.paymentSuccess(event);
+		paymentEventHandler.paymentSuccess(event);
 	}
 
 	@EventListener
 	public void handlePaymentSuccessEvent(PaymentEvent.PaymentFail event) {
-		paymentFacade.paymentFail(event);
+		paymentEventHandler.paymentFail(event);
 	}
 }
